+++ lib/new_relic/harvest/collector/custom_event/harvester.ex
@@ -31,7 +31,7 @@ defmodule NewRelic.Harvest.Collector.CustomEvent.Harvester do
     do:
       %Event{
         type: type,
-        attributes: annotate(attributes),
+        attributes: process(attributes),
         timestamp: System.system_time(:millisecond) / 1_000
       }
       |> report_custom_event
@@ -74,8 +74,9 @@ defmodule NewRelic.Harvest.Collector.CustomEvent.Harvester do
 
   # Helpers
 
-  def annotate(event) do
+  def process(event) do
     event
+    |> NewRelic.Util.coerce_attributes()
     |> Map.merge(NewRelic.Config.automatic_attributes())
   end