+++ test/aggregate_test.exs
@@ -44,7 +44,7 @@ defmodule AggregateTest do
     events = TestHelper.gather_harvest(Collector.CustomEvent.Harvester)
 
     assert Enum.find(events, fn [_, event, _] ->
-             event[:category] == :Metric && event[:meta] == "data" && event[:duration] == 15
+             event[:category] == "Metric" && event[:meta] == "data" && event[:duration] == 15
            end)
   end
 end